% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_calendarevent.R
\name{calendarevent}
\alias{calendarevent}
\alias{suso_createCALEV}
\alias{suso_updateCALEV}
\alias{suso_delCALEV}
\title{Create, update or delete a calendar event}
\usage{
suso_createCALEV(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = NULL,
  AssId = NULL,
  comment = NULL,
  startDate = NULL,
  startTime = NULL,
  startTZ = "UTC"
)

suso_updateCALEV(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  publicKey = NULL,
  comment = NULL,
  startDate = NULL,
  startTime = NULL,
  startTZ = "UTC"
)

suso_delCALEV(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  publicKey = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and
\emph{apiPass} will be ignored}

\item{intID}{UUID vector of interview id(s)}

\item{AssId}{numeric vector of assignment id(s)}

\item{comment}{a comment string}

\item{startDate}{new start date, format must be: \code{2024-01-16}}

\item{startTime}{new start date, format must be: \code{01:41:14}}

\item{startTZ}{time zone of the tablet device, use \code{\link[base]{OlsonNames}}}

\item{publicKey}{UUID vector of calender event id(s)}
}
\value{
a data.table with the created event(s).
}
\description{
Create, update or delete a calendar event for one or several assignments or interviews.
}
\details{
This function creates a calendar event either for one or several assignments or one or
several interviews. Either one of \code{AssID} or \code{intID} must not be NULL. If
\code{length(AssID)} or \code{length(AssID)} is greater 1, then \code{startDate},
\code{startTime} and \code{startTZ} must either be each of length 1 (same event for
all) or of the same length, allowing for multiple events with different times.
}
\section{Functions}{
\itemize{
\item \code{suso_createCALEV()}: add a calendar event

\item \code{suso_updateCALEV()}: update a calendar event

\item \code{suso_delCALEV()}: delete a calendar event

}}
\examples{
\dontshow{if (suso_PwCheck()==200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# add single calendar event to assignment
calEvass<-suso_createCALEV(AssId = 268,
                           startDate = "2024-01-17",
                           startTime = "10:40:00")

# update single event
calEvup<-suso_updateCALEV(publicKey = calEvass$publicKey[1],
                         startDate = "2024-01-18",
                        startTime = "10:40:00")

# delete single event
calEvDel<-suso_delCALEV(publicKey = calEvup$publicKey[1])


# add calendar event to multiple assignments
calEvass<-suso_createCALEV(AssId = 55:60,
                           startDate = "2024-01-18",
                           startTime = "11:40:00")
# delete multiple calendar events
calEvDel<-suso_delCALEV(publicKey = calEvass$publicKey)

\dontshow{\}) # examplesIf}
}
