% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.suso_transform_fullValid_q}
\alias{.suso_transform_fullValid_q}
\title{Helper functions to transform list of questionnair structure to single data.table with all variables}
\usage{
.suso_transform_fullValid_q(input = NULL)
}
\arguments{
\item{input}{returned by \code{suso_getQuestDetails} structure operation}
}
\value{
data.table with all variables in the questionnaire
}
\description{
Uses tidyjson package
}
\details{
\code{suso_transform_fullMeta} transforms the list containing the structure (\emph{operation.type = structure})
is transformed into a single data.table
with all variable names, types etc.. This also works with json strings manually exported from the server.
}
\keyword{internal}
