% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_assignment_class.R
\name{is.assignmentClass}
\alias{is.assignmentClass}
\title{Function to check if an object is of class assignmentClass}
\usage{
is.assignmentClass(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
TRUE if object is of class assignmentClass
}
\description{
Function to check if an object is of class assignmentClass
}
\examples{
ac <- assignmentClass(list(Assignments = data.frame(Id = 1)))
is.assignmentClass(ac)
is.assignmentClass(data.frame())
}
