% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_export_class.R
\name{summaryTable}
\alias{summaryTable}
\title{S3 method to create a summary table of numeric variables for an exportClass object}
\usage{
summaryTable(x, includeFactors = FALSE, useDT = TRUE, DTstyle = TRUE, ...)
}
\arguments{
\item{x}{A exportClass object.}

\item{includeFactors}{if \code{TRUE}, factor variables will be converted to numeric and included in the table.}

\item{useDT}{Logical, if TRUE (default) the function returns a DataTable (DT) object (HTML table), otherwise
a data.table object is returned.}

\item{DTstyle}{if \code{TRUE} and \code{useDT = TRUE}, then a custom style will be applied, otherwise a plain DT table will be returned.}

\item{...}{Additional arguments (not used).}
}
\value{
A DataTable (DT) object (HTML table) displaying the summary statistics, if \code{useDT = TRUE}, and a plain data.table otherwise.
}
\description{
This function generates a summary table of numeric variables in an exportClass
object. The table includes statistics such as mean, standard deviation, maximum,
minimum, and total count as well as count of NA values.
}
\examples{
\dontrun{
# Load your package
export <- suso_export(questID = questlist$QuestionnaireId[7],
                      version = questlist$Version[7],
                      combineFiles = T,
                      reloadTimeDiff = 20)

# Create a summary table
summary_data_table <- summaryTable(export)

# Display the summary table
summary_data_table
}

}
