% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_user_class.R
\name{summaryTable.UserClass}
\alias{summaryTable.UserClass}
\title{S3 method to create a summary table of users for UserClass object}
\usage{
\method{summaryTable}{UserClass}(x, includeFactors = FALSE, useDT = TRUE, DTstyle = TRUE, ...)
}
\arguments{
\item{x}{A UserClass object.}

\item{includeFactors}{if \code{TRUE}, factor variables will be converted to numeric and included in the table (only for exportClass).}

\item{useDT}{Logical, if TRUE (default) the function returns a DataTable (DT) object (HTML table), otherwise
a data.table object is returned.}

\item{DTstyle}{if \code{TRUE} and \code{useDT = TRUE}, then a custom style will be applied, otherwise a plain DT table will be returned.}

\item{...}{Additional arguments (not used).}
}
\value{
A DataTable (DT) object (HTML table)  if \code{useDT = TRUE}, and a plain data.table otherwise. The table contains
all users core data, plus completed and assigned counts,
}
\description{
This function generates a summary table of Users.
}
\details{
For very large tables you will get a warning when rendering in Rstudio, however in a shiny app you have the option to use
server side processing: \code{DT::renderDataTable(..., server = TRUE)}. In cases where \code{future} and \code{future.apply} is
available on your system the additional information will be retrieved in parallel (multisession), otherwise sequential. For large
data sets parallel processing is recommended. Furthermore when processing in parallel,
a progress bar for shiny applications is included and active when running in a shiny application, otherwise a cli progress bar will
be displayed.
}
\examples{
\dontrun{

allintprim<-suso_getINT(workspace = "primary")

# Create a summary table
summary_data_table <- summaryTable(allintprim)

# Display the summary table
summary_data_table
}

}
