% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_assignInterview}
\alias{suso_assignInterview}
\title{Assign interview to interviewer or supervisor}
\usage{
suso_assignInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = "",
  userId = NULL,
  userName = NULL,
  supervisor = FALSE
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace name}

\item{token}{API token}

\item{intID}{InterviewId (GUID)}

\item{userId}{User GUID of responsible person}

\item{userName}{Name of responsible person (optional)}

\item{supervisor}{If TRUE assigns to a supervisor, otherwise assigns to an interviewer}
}
\value{
A data.table indicating the assignment status.
}
\description{
Reassigns an interview to a specified interviewer or supervisor.
}
\examples{
\dontrun{
suso_assignInterview(
  intID = "11111111-2222-3333-4444-555555555555",
  userName = "interviewer01"
)
}
}
