% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_export_data.R
\name{suso_cancelExport}
\alias{suso_cancelExport}
\title{Survey Solutions API call to cancel an export process}
\usage{
suso_cancelExport(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = suso_get_api_key("workspace"),
  jobid = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{jobid}{Export process ID (integer)}
}
\value{
A data.table containing the canceled export process information.
}
\description{
Cancels an ongoing export process or deletes an export process by ID.
}
\examples{
\dontrun{
# Cancel an ongoing export process
suso_cancelExport(jobid = 123)
}
}
