% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_commentInterview}
\alias{suso_commentInterview}
\title{Leave a comment on a question in an interview}
\usage{
suso_commentInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = "",
  comment = "",
  questionId = NULL,
  variable = NULL,
  rosterVector = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace name}

\item{token}{API token}

\item{intID}{InterviewId (GUID)}

\item{comment}{The comment text to leave}

\item{questionId}{The question GUID (optional if variable is provided)}

\item{variable}{Variable name of question (optional if questionId is provided)}

\item{rosterVector}{Integer vector specifying roster indices (e.g. c(0, 1))}
}
\value{
A data.table indicating the comment status.
}
\description{
Leaves a comment on a question using either the questionId or the questionnaire variable name.
}
\examples{
\dontrun{
suso_commentInterview(
  intID = "11111111-2222-3333-4444-555555555555",
  variable = "age",
  comment = "Please verify respondent age."
)
}
}
