% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_create_user.R
\name{suso_createUSER}
\alias{suso_createUSER}
\title{Survey Solutions API call for User Creation}
\usage{
suso_createUSER(
  userlist = NULL,
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  showUser = FALSE
)
}
\arguments{
\item{userlist}{dataframe with upload data}

\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{showUser}{userName in console}
}
\value{
A data.table containing the user information and HTTP response status code for each created user.
}
\description{
Creates Survey Solutions users (observers, interviewers or supervisors).
}
\details{
Dataframe needs to be provided with the mandatory columns
for user creation which are: Role, UserName, Password and Supervisor (in case of interviewer),
optional you can also provide FullName, PhoneNumber and Email. Return value is a data.table, which includes
the user information as well as the response's status code. Important is also that the UserName and Password are
provided in the required format.
}
\examples{
\dontrun{
new_users <- data.frame(
  Role = "Interviewer",
  UserName = "interviewer01",
  Password = "Password123!",
  Supervisor = "supervisor01"
)
suso_createUSER(userlist = new_users)
}

}
