% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_workspace.R
\name{suso_createWorkspace}
\alias{suso_createWorkspace}
\title{Survey Solutions API call to create workspace}
\usage{
suso_createWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  new_workspace = NULL,
  displayName = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions ADMIN user}

\item{apiPass}{Survey Solutions ADMIN password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{new_workspace}{The name used by the system for this workspace. Make sure you follow the nameing rules outlined
under \url{https://docs.mysurvey.solutions/headquarters/accounts/workspaces/}}

\item{displayName}{The name visible to the users}
}
\value{
A data.table containing the created workspace information.
}
\description{
\code{suso_createWorkspace} Allows you to create a workspace with a specific system name and display name.
For more details please read \url{https://docs.mysurvey.solutions/headquarters/accounts/workspaces/}. To run this command
you require admin credentials.
}
\details{
Be aware, that for using this call you require the ADMIN credentials, and not the regular API user credentials.
}
\examples{
\dontrun{
# Use Admin Credentials!
suso_createWorkspace(
          new_workspace = "myworkspace1",
          displayName = "SpecialWorkspace",
          apiUser = "xxxxxx",
          apiPass = "xxxxxx")
}


}
