% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_deleteInterview}
\alias{suso_deleteInterview}
\title{Delete an interview}
\usage{
suso_deleteInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = ""
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace name}

\item{token}{API token}

\item{intID}{InterviewId of the interview (GUID)}
}
\value{
A data.table indicating the deletion status.
}
\description{
Deletes an interview from the Survey Solutions server.
}
\examples{
\dontrun{
suso_deleteInterview(intID = "11111111-2222-3333-4444-555555555555")
}
}
