% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_maps.R
\name{suso_deletemap}
\alias{suso_deletemap}
\title{Delete Map from Server}
\usage{
suso_deletemap(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  fileName = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{fileName}{the name of the map file on the server}
}
\value{
A data.table containing the deleted map details (file name, shape type, and import timestamp).
}
\description{
Allows the user to delete maps stored on the server.
}
\examples{
\dontrun{
suso_deletemap(
  workspace = "myworkspace",
  fileName = "Lat9264Lon625_ALL.tif"
)
}

}
