% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_workspace.R
\name{suso_deleteWorkspace}
\alias{suso_deleteWorkspace}
\title{Delete a workspace}
\usage{
suso_deleteWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{API token}

\item{workspace}{Name of the workspace to delete}
}
\value{
A data.table indicating the deletion status.
}
\description{
Deletes a specified workspace. Accessible only to administrator.
}
\examples{
\dontrun{
suso_deleteWorkspace(workspace = "old_ws")
}
}
