% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_workspace.R
\name{suso_enableWorkspace}
\alias{suso_enableWorkspace}
\title{Enable or disable a workspace}
\usage{
suso_enableWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = NULL,
  enable = TRUE
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{API token}

\item{workspace}{Name of the workspace}

\item{enable}{Logical, if TRUE enables the workspace, if FALSE disables it. Default is TRUE.}
}
\value{
A data.table indicating the enable/disable status.
}
\description{
Enables or disables access to a specified workspace. Accessible only to administrator.
}
\examples{
\dontrun{
# Enable workspace
suso_enableWorkspace(workspace = "my_ws", enable = TRUE)
# Disable workspace
suso_enableWorkspace(workspace = "my_ws", enable = FALSE)
}
}
