% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_credentials.R
\name{suso_get_default_key}
\alias{suso_get_default_key}
\title{Checks if credentials are present}
\usage{
suso_get_default_key(
  api = c("susoServer", "susoUser", "susoPass", "workspace")
)
}
\arguments{
\item{api}{one of susoServer, susoUser, susoPass, or workspace}
}
\value{
Character string of the requested credential, or throws error if not set.
}
\description{
Helper function
}
\examples{
\dontrun{
suso_get_default_key("susoServer")
}
}
