% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_statistics.R
\name{suso_get_stats}
\alias{suso_get_stats}
\title{Survey Solutions API call for Summary Tables}
\usage{
suso_get_stats(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  questID = NULL,
  version = NULL,
  qQuest = "",
  byTeam = TRUE,
  exportType = c("Csv", "Json")
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{Questionnaire ID (GUID)}

\item{version}{Version of the questionnaire (numeric)}

\item{qQuest}{Question variable name or UUID}

\item{byTeam}{Logical, whether the table should contain reports by team (default: \code{TRUE})}

\item{exportType}{Report format: \code{"Csv"} (default) or \code{"Json"}}
}
\value{
A data.table containing the summary statistics for the question.
}
\description{
Returns summary tables for individual questions from Survey Solutions statistics.
If no responses have been provided, an empty table will be returned.
}
\examples{
\dontrun{
suso_get_stats(
  questID = "11111111-2222-3333-4444-555555555555",
  version = 1,
  qQuest = "age"
)
}
}
