% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_get_stats_interview}
\alias{suso_get_stats_interview}
\title{Get statistics for interview}
\usage{
suso_get_stats_interview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = ""
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{intID}{a single or multiple \emph{InterviewId}.}
}
\value{
A data.table containing statistics for the requested interview(s).
}
\description{
Fetch statistics for a specific interview, vectorized over interview id (int_id)
}
\examples{
\dontrun{
suso_get_stats_interview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea"
          )

}

}
