% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_export_data.R
\name{suso_getExportList}
\alias{suso_getExportList}
\title{Survey Solutions API call to list export processes}
\usage{
suso_getExportList(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = suso_get_api_key("workspace"),
  exportType = NULL,
  interviewStatus = NULL,
  questID = NULL,
  version = NULL,
  exportStatus = NULL,
  hasFile = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{exportType}{Format of export data: \code{"Tabular"}, \code{"STATA"}, \code{"SPSS"}, \code{"Binary"}, \code{"DDI"}, \code{"Parquet"}}

\item{interviewStatus}{Status of exported interviews: \code{"All"}, \code{"SupervisorAssigned"}, \code{"InterviewerAssigned"},
\code{"RejectedBySupervisor"}, \code{"Completed"}, \code{"ApprovedBySupervisor"}, \code{"RejectedByHeadquarters"}, \code{"ApprovedByHeadquarters"}}

\item{questID}{Questionnaire ID (GUID)}

\item{version}{Questionnaire version}

\item{exportStatus}{Status of export process: \code{"Created"}, \code{"Running"}, \code{"Completed"}, \code{"Fail"}, \code{"Canceled"}}

\item{hasFile}{Logical, whether the export process has a file ready to download}

\item{limit}{Maximum number of records to return}

\item{offset}{Number of records to skip}
}
\value{
A data.table containing the list of export processes.
}
\description{
Retrieves a list of export processes matching specified filters.
}
\examples{
\dontrun{
# List all export processes in workspace
exp_list <- suso_getExportList()
}
}
