% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_export_data.R
\name{suso_getExportProcess}
\alias{suso_getExportProcess}
\title{Survey Solutions API call to get detailed information about an export process}
\usage{
suso_getExportProcess(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = suso_get_api_key("workspace"),
  jobid = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{jobid}{Export process ID (integer)}
}
\value{
A data.table containing export process details.
}
\description{
Retrieves status and details of a single export process by its ID.
}
\examples{
\dontrun{
# Get status of export process by jobid
proc_info <- suso_getExportProcess(jobid = 123)
}
}
