% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Questionnaire.R
\name{suso_getQuestDetails}
\alias{suso_getQuestDetails}
\title{Survey Solutions API call for questionnaire}
\usage{
suso_getQuestDetails(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  questID = NULL,
  version = NULL,
  operation.type = c("list", "statuses", "structure", "interviews"),
  include_raw = FALSE,
  AssId = NULL,
  InterviewKey = NULL,
  errorsCount = NULL,
  errosCountFilter = c("lower", "higher", "equal"),
  interviewMode = c("CAPI", "CAWI"),
  notAnsweredCount = NULL,
  notAnsweredCountFilter = c("lower", "higher", "equal"),
  QuestionnaireVariable = NULL,
  ResponsibleName = NULL,
  responsibleRole = c("INTERVIEWER", "SUPERVISOR"),
  workStatus = c("All", "SupervisorAssigned", "InterviewerAssigned",
    "RejectedBySupervisor", "Completed", "ApprovedBySupervisor",
    "RejectedByHeadquarters", "ApprovedByHeadquarters"),
  supervisorName = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}

\item{questID}{\emph{QuestionnaireId} for which details should be exported}

\item{version}{questionnaire version}

\item{operation.type}{if \emph{list} is specified a list of all questionnaires on the server. If
\emph{statuses} a vector of all questionnaire statuses. If \emph{structure} is specified, it returns a list
containing all questions, rosters etc. of the specific questionnaire, as well as all validations.
If \emph{interviews} is specified, all interviews for a specific questionnaire. See details bellow.}

\item{include_raw}{logical; if \code{TRUE}, retains the raw JSON list column (\code{..JSON}) in the question table when \code{operation.type = "structure"}. Default is \code{FALSE} for clean, flat data tables.}

\item{AssId}{Assignment ID (only required if operations.type is\emph{interviews})}

\item{InterviewKey}{Interview key (only required if operations.type is\emph{interviews})}

\item{errorsCount}{desired number of errors (only required if operations.type is\emph{interviews})}

\item{errosCountFilter}{relational type for error counts, either smaller equal, equal or greater equal than
the number specified in \code{errorsCount}, if not supplied defaults to lower equal than  \emph{interviews})
(only required if operations.type is\emph{interviews})}

\item{interviewMode}{Interview mode (CAWI or CAPI) (only required if operations.type is\emph{interviews})}

\item{notAnsweredCount}{number of unanswered questions (only required if operations.type is\emph{interviews})}

\item{notAnsweredCountFilter}{relational type for unanswered question counts, either smaller equal, equal or greater equal than
the number specified in \code{notAnsweredCount}, if not supplied defaults to lower equal than  \emph{interviews})
(only required if operations.type is\emph{interviews})}

\item{QuestionnaireVariable}{the variable for the questionnaire (only required if operations.type is\emph{interviews})}

\item{ResponsibleName}{Name of the person responsible (only required if operations.type is\emph{interviews})}

\item{responsibleRole}{Role of the person responsible (only required if operations.type is\emph{interviews})}

\item{workStatus}{of the interview (only required if operations.type is\emph{interviews})}

\item{supervisorName}{Name of the supervisor of the responsible user (only required if operations.type is\emph{interviews})}
}
\value{
Depending on \code{operation.type}:
\describe{
\item{list}{A data.table listing all questionnaires on the server.}
\item{statuses}{A character vector of questionnaire statuses.}
\item{structure}{A list with four data.tables: \code{q} (questions/rosters metadata), \code{val} (validations), \code{v} (alias for \code{val}), and \code{answers} (categorical answer options and codes).}
\item{interviews}{A data.table listing interviews for the specified questionnaire.}
}
}
\description{
\code{suso_getQuestDetails} implements all Questionnaire related API commands. It allows for different operation types,
see details bellow for further clarification.
}
\details{
If list is selected, then list of questionnaires is returned.

If statuses is selected, a list of all available questionnaire statuses is returned (deprecated).

In case structure is chosen the return value is a list with four data.table elements:
\itemize{
\item List element \emph{q} contains all questions, rosters etc. with full metadata.
\item List element \emph{val} contains all validations.
\item List element \emph{v} is an alias for \emph{val} for backward compatibility.
\item List element \emph{answers} contains categorical answer options, codes, and linked questions.
}
In this way it is straightforward to use the return value for questionnaire manuals and the likes.

In case interviews is selected, a list of all interviews for the specific questionnaire is returned.
}
\examples{
\dontrun{
# List all questionnaires on the server
q_list <- suso_getQuestDetails(operation.type = "list")

# Get questionnaire structure (questions, validations, answers)
q_struct <- suso_getQuestDetails(
  questID = q_list$QuestionnaireId[1],
  version = q_list$Version[1],
  operation.type = "structure"
)
}

}
