% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_statistics.R
\name{suso_getQuestionsQuestionnaire}
\alias{suso_getQuestionsQuestionnaire}
\title{Survey Solutions API call for questions and responses from single questionnaire}
\usage{
suso_getQuestionsQuestionnaire(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  questID = NULL,
  version = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{Questionnaire ID (GUID)}

\item{version}{Questionnaire version (numeric)}
}
\value{
A data.table listing questions with responses for the questionnaire.
}
\description{
Returns all questions for a single questionnaire (ONLY if they contain responses). If you require all questions
from any questionnaire on the server, use \code{suso_getQuestDetails(..., operation.type = "structure")}.
}
\examples{
\dontrun{
suso_getQuestionsQuestionnaire(
  questID = "11111111-2222-3333-4444-555555555555",
  version = 1
)
}
}
