% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_statistics.R
\name{suso_getStatsQuestionnaires}
\alias{suso_getStatsQuestionnaires}
\title{Survey Solutions API call for questionnaires with statistics}
\usage{
suso_getStatsQuestionnaires(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}
}
\value{
A data.table listing questionnaires with statistics.
}
\description{
Retrieves the list of questionnaires that have statistics data on the server.
}
\examples{
\dontrun{
suso_getStatsQuestionnaires()
}
}
