% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_user.R
\name{suso_getSV}
\alias{suso_getSV}
\title{Survey Solutions API call for list of supervisors}
\usage{
suso_getSV(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}
}
\value{
An object of class \code{UserClass} (inheriting from data.table) containing the list of supervisors.
}
\description{
Gets list of supervisors
}
\examples{
\dontrun{
# receive all supervisors in the workspace
suso_getSV()
}

}
