% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_user.R
\name{suso_getSV_info}
\alias{suso_getSV_info}
\title{Survey Solutions API call for supervisor info}
\usage{
suso_getSV_info(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  sv_id = NULL,
  workspace = suso_get_api_key("workspace"),
  token = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{sv_id}{supervisor ID (GUID)}

\item{workspace}{server workspace name}

\item{token}{API token}
}
\value{
A data.table containing supervisor details.
}
\description{
Gets detailed info about a single supervisor by supervisor ID.
}
\examples{
\dontrun{
suso_getSV_info(
  workspace = "myworkspace",
  sv_id = "xxxx-xxxx-xxxx-xxx"
)
}

}
