% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_workspace.R
\name{suso_getWorkspace}
\alias{suso_getWorkspace}
\title{Survey Solutions API call for workspace information}
\usage{
suso_getWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = NULL,
  status = FALSE
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace}

\item{status}{if status is \emph{TRUE} worskpace must be not NULL and status information about a specific
workspace is requested}
}
\value{
A data.table containing workspace information or status.
}
\description{
\code{suso_getWorkspace} allows you to get the list of workspaces, information about individual workspace names
as well as workspace statuses. Workspaces as well as Workspaces information can only be accessed, if credentials
are eligible. For more details please read \url{https://docs.mysurvey.solutions/headquarters/accounts/workspaces/}
}
\examples{
\dontrun{
# This assumes, that suso_PwCheck(workspace = "myworkspace") was
# sucessful

# shows all workspaces in the system AND the user has access to
suso_getWorkspace(
          status = F)

# shows details for specific workspace myworkspace
suso_getWorkspace(
          workspace = "myworkspace",
          status = T)
}

}
