% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_settings.R
\name{suso_globalNotice}
\alias{suso_globalNotice}
\title{Survey Solutions API call for Headquarters Global Notice settings}
\usage{
suso_globalNotice(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  action = c("get", "set", "delete"),
  message = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{action}{Action to perform: \code{"get"} to retrieve the notice, \code{"set"} to update the notice, or \code{"delete"} to remove the notice}

\item{message}{Character string containing the notice message. Required when \code{action = "set"}.}
}
\value{
When \code{action = "get"}, returns a character string with the current notice (or NULL if no notice is set).
When \code{action = "set"} or \code{"delete"}, returns \code{TRUE} invisibly on success.
}
\description{
Allows getting, setting, and removing the global notice displayed in the Survey Solutions Headquarters application.
}
\examples{
\dontrun{
# Get current global notice
suso_globalNotice(action = "get")

# Set a new global notice
suso_globalNotice(action = "set",
                  message = "Maintenance tonight at 22:00 UTC.")

# Delete the global notice
suso_globalNotice(action = "delete")
}

}
