% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_maps.R
\name{suso_mapassign}
\alias{suso_mapassign}
\title{Assigns/Unassign a map to a user}
\usage{
suso_mapassign(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  fileName = NULL,
  userName = NULL,
  assignUser = TRUE
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{fileName}{the name of the map file on the server}

\item{userName}{the name of the interviewer to whom the map will be assigned to}

\item{assignUser}{if TRUE user will be assigned to map, if FALSE user will be removed from map}
}
\value{
A data.table containing the map assignment details (file name, user name, shape type, and import timestamp).
}
\description{
Allows the user to assign/unassign a map to an interviewer to be used in CAPI data collection.
}
\examples{
\dontrun{
suso_mapassign(workspace = "myworkspace",
              fileName = "Lat9264Lon625_ALL.tif",
              userName = "INT0004")
}

}
