% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_maps.R
\name{suso_mapreport}
\alias{suso_mapreport}
\title{Create a Map Report}
\usage{
suso_mapreport(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  questID = NULL,
  version = NULL,
  variable = NULL,
  zoom = 1,
  clientMapWidth = 0,
  west = -180,
  east = 180,
  north = 90,
  south = -90,
  assignmentId = NULL,
  clientKey = NULL,
  createdDate = NULL,
  errorsCount = NULL,
  identifyingData = NULL,
  interviewMode = NULL,
  notAnsweredCount = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{variable}{Variable(s) of interest}

\item{zoom}{Zoom of the map report}

\item{clientMapWidth}{width of the client map}

\item{west}{coordinates for bounding box}

\item{east}{coordinates for bounding box}

\item{north}{coordinates for bounding box}

\item{south}{coordinates for bounding box}

\item{assignmentId}{Assignment ID}

\item{clientKey}{Interview key}

\item{createdDate}{Creation data of the interview}

\item{errorsCount}{number of errors}

\item{identifyingData}{Pre-loaded identifying data}

\item{interviewMode}{Interview mode (CAWI or CAPI)}

\item{notAnsweredCount}{number of unanswered questions}
}
\value{
A list containing the map report result.
}
\description{
Allows the user to create a map report with pre-specified parameters.
}
\examples{
\dontrun{
suso_mapreport(
  workspace = "myworkspace",
  questID = "11111111-2222-3333-4444-555555555555",
  version = 1,
  variable = "gps_loc"
)
}

}
