% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_patchApproveInterview}
\alias{suso_patchApproveInterview}
\title{Approve interviews either as supervisor or as headquarter.}
\usage{
suso_patchApproveInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = "",
  HQ = FALSE,
  hqunapprove = FALSE,
  comment = "Well done!"
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{intID}{the \emph{InterviewId} of the interview.}

\item{HQ}{if FALSE, approve as supervisor, if TRUE approve as headquarters}

\item{hqunapprove}{if TRUE, unapprove from headquarters status (hqunapprove)}

\item{comment}{comment which should be sent with the questionnaire}
}
\value{
A data.table containing the status of the approval operation.
}
\description{
Allows you to approve interviews in supervisor or headquarters role,
unapprove from headquarters status, as well as to provide a comment (i.e. reason).
}
\details{
For details please see:
\url{https://docs.mysurvey.solutions/headquarters/interviews/survey-workflow/}
}
\examples{
\dontrun{
# approve the interview as supervisor
suso_patchApproveInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea"
          )
# approve the interview as headquarters
suso_patchApproveInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
          HQ = TRUE
          )
}

}
