% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_credentials.R
\name{suso_set_workspace}
\alias{suso_set_workspace}
\title{Set workspace only}
\usage{
suso_set_workspace(workspace = NULL)
}
\arguments{
\item{workspace}{server workspace, if nothing provided, defaults to primary}
}
\description{
Sets the workspace only, but leaves all other credentials the same.
}
\details{
Use \code{suso_set_workspace} to make the desired workspace available for all the \code{suso_}
functions, so you don't need to specify the workspace parameter within those
functions.
}
