% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_workspace.R
\name{suso_updateWorkspace}
\alias{suso_updateWorkspace}
\title{Update workspace details}
\usage{
suso_updateWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  workspace = NULL,
  displayName = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{token}{API token}

\item{workspace}{Name of the workspace to update}

\item{displayName}{New display name for the workspace}
}
\value{
A data.table indicating the update status.
}
\description{
Updates the display name of a workspace. Accessible only to administrator.
}
\examples{
\dontrun{
suso_updateWorkspace(workspace = "my_ws", displayName = "Renamed Workspace")
}
}
