% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_user_class.R
\name{UserClass}
\alias{UserClass}
\title{User Classes & Methods}
\usage{
UserClass(x, ...)
}
\arguments{
\item{x}{list returned by api call}

\item{...}{additional attributes to be added to the assignmentClass}
}
\value{
An object of class UserClass (inheriting from data.table)
}
\description{
Users class extends the data.table class to include relevant methods
}
\examples{
usr_list <- list(Users = data.frame(UserName = "interviewer1", Role = "Interviewer"))
uc <- UserClass(usr_list)
is.UserClass(uc)
}
