% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_export_class.R
\name{boxplot_summary}
\alias{boxplot_summary}
\title{S3 method to create a box plot of numeric variables for an exportClass object}
\usage{
boxplot_summary(x, useGGplot2 = FALSE, ...)
}
\arguments{
\item{x}{A summaryTable object.}

\item{useGGplot2}{Logical, if TRUE (default) the function returns a ggplot2 object, otherwise
a base R boxplot is returned.}

\item{...}{Additional arguments (not used).}
}
\value{
A box plot visualizing the numeric variables.
}
\description{
This function generates a box plot for the numeric variables in a exportClass
object.
}
\examples{
\dontrun{
# Load your package
export <- suso_export(questID = questlist$QuestionnaireId[7],
                      version = questlist$Version[7],
                      combineFiles = T,
                      reloadTimeDiff = 20)

# Create a summary table
boxplot_summary <- boxplot(export)

# Display the summary table
boxplot_summary
}

}
