% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_export_class.R
\name{exportClass}
\alias{exportClass}
\title{Export Classes & Methods}
\usage{
exportClass(x, varLabels, ..., type = "main")
}
\arguments{
\item{x}{list returned by api call}

\item{varLabels}{data.table of variable labels as returned by
\code{suso_getQuestDetails(operation.type = "structure")}}

\item{...}{additional attributes to be added to the ExportClass}

\item{type}{one of main or para, if main returns exportClass object for the main data,
if para, returns an exportClass object for the paradata.}
}
\description{
Export class extends the data.table class to include relevant methods
}
