% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_assignment_class.R, R/as_export_class.R,
%   R/as_user_class.R
\name{getinfo}
\alias{getinfo}
\title{assignmentClass methods}
\usage{
getinfo(obj, arg)

getinfo(obj, arg)

getinfo(obj, arg)
}
\arguments{
\item{obj}{object of UserClass}

\item{arg}{name of attribute, if \code{arg="arglist"} then it returns all available arguments}
}
\value{
the specific attribute

the specific attribute

the specific attribute
}
\description{
\code{getinfo} allows you to retrieve relevant additional information from the \code{assignmentClass},
object depending on the api endpoint.

\code{getinfo} allows you to retrieve relevant additional information from the \code{ExportClass},
object depending on the api endpoint.

\code{getinfo} allows you to retrieve relevant additional information from the \code{UserClass},
object depending on the api endpoint.
}
\details{
To retrieve all availalbe arguments use \code{arg="arglist"}

To retrieve all availalbe arguments use \code{arg="arglist"}

To retrieve all availalbe arguments use \code{arg="arglist"}
}
\examples{
\dontrun{

# retrieve the uid of the person responsible after retrieving details for specific assignment
getinfo(asslist, "responsibleid")

# see all available attribute names
getinfo(asslist, "arglist")

}

\dontrun{

# retrieve the uid of the person responsible after retrieving details for specific User
getinfo(asslist, "responsibleid")

# see all available attribute names
getinfo(asslist, "arglist")

}

\dontrun{

# retrieve the uid of the person responsible after retrieving details for specific User
getinfo(asslist, "responsibleid")

# see all available attribute names
getinfo(asslist, "arglist")

}

}
