% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_export_class.R
\name{is.exportClass}
\alias{is.exportClass}
\title{Function to check if an object is of class exportClass}
\usage{
is.exportClass(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
TRUE if object is of class exportClass
}
\description{
Function to check if an object is of class exportClass
}
