% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_user_class.R
\name{is.UserClass}
\alias{is.UserClass}
\title{Function to check if an object is of class UserClass}
\usage{
is.UserClass(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
TRUE if object is of class UserClass
}
\description{
Function to check if an object is of class UserClass
}
