% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_export_class.R
\name{summaryTable}
\alias{summaryTable}
\title{S3 method to create a summary table of numeric variables for an exportClass object}
\usage{
summaryTable(x, useDT = TRUE, ...)
}
\arguments{
\item{x}{A exportClass object.}

\item{useDT}{Logical, if TRUE (default) the function returns a DataTable (DT) object (HTML table), otherwise
a data.table object is returned.}

\item{...}{Additional arguments (not used).}
}
\value{
A DataTable (DT) object (HTML table) displaying the summary statistics.
}
\description{
This function generates a summary table of numeric variables in a exportClass
object. The table includes statistics such as mean, standard deviation, maximum,
minimum, and count of NA values.
}
\examples{
\dontrun{
# Load your package
export <- suso_export(questID = questlist$QuestionnaireId[7],
                      version = questlist$Version[7],
                      combineFiles = T,
                      reloadTimeDiff = 20)

# Create a summary table
summary_data_table <- summaryTable(export)

# Display the summary table
summary_data_table
}

}
