% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySolutionsAPIv2-package.R
\docType{package}
\name{SurveySolutionsAPIv2-package}
\alias{SurveySolutionsAPIv2}
\alias{SurveySolutionsAPIv2-package}
\title{SurveySolutionsAPIv2: Comprehensive API access to your Survey Solutions Server (httr2 version)}
\description{
This packages implements all the currently available POST, GET, and PATCH calls to the Survey Solutions RESTfull API. It contains several convenience functions to facilitate these operations, and can handle small impact evaluations as well as large scale cenuses. It is based on the httr2 package, and is officially replacing the httr based version.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
