% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_options.R
\name{suso_set_pwcheck_mess}
\alias{suso_set_pwcheck_mess}
\alias{suso_set_prog_mess}
\title{Convenience functions to modify user notification text}
\usage{
suso_set_pwcheck_mess(mess_succ = NULL, mess_fail = NULL)

suso_set_prog_mess(mess = NULL)
}
\arguments{
\item{mess_succ}{set the success message for \code{\link{suso_PwCheck}} when used in shiny app.}

\item{mess_fail}{set the success message for \code{\link{suso_PwCheck}} when used in shiny app.}

\item{mess}{set the message for the export progress bars when used in shiny app, used in \code{\link{suso_export}} and \code{\link{suso_export_paradata}}}
}
\value{
always invisible \code{TRUE}.
}
\description{
These functions allow the user to customize certain notifications and messages when used in shiny apps.
}
\concept{suso-customization-options}
