% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_options.R
\name{suso.options}
\alias{suso.options}
\title{Configuration options used for the SurveySolutionsAPIv2 package}
\description{
Below are the options and environment variables, that are used by the \pkg{SurveySolutionsAPIv2} package.
}
\section{Available options}{

Bellow is a list of options, with their default values.

\itemize{
\item \option{suso.maxpar.req} specifies the maximum number of parallel requests, default is 100
\item \option{suso.maxpar.con} specifies the maximum number of parallel connections,
and must always be equal or greater than suso.maxpar.req, default is 100.
}
}

