% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_user.R
\name{suso_archUSR}
\alias{suso_archUSR}
\title{Survey Solutions API call (un-) archive user}
\usage{
suso_archUSR(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  user_id = NULL,
  archive = F,
  workspace = suso_get_api_key("workspace"),
  token = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{user_id}{user id}

\item{archive}{if TRUE user will be archived or statys archived, if FALSE user will be unarchived or stays unarchived}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}
}
\description{
(Un-)Archive user
}
\examples{
\dontrun{
# you can archive a user by archive=T
suso_archUSR(
          workspace = "myworkspace",
          uid = "xxxx-xxxx-xxxx-xxx",
          archive = TRUE
          )
# and unarchive a user by archive=F
suso_archUSR(
          workspace = "myworkspace",
          uid = "xxxx-xxxx-xxxx-xxx",
          archive = FALSE
          )
}

}
