% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_create.R
\name{suso_createASS}
\alias{suso_createASS}
\title{Survey Solutions API call for Assignment Creation (httr2 version)}
\usage{
suso_createASS(
  df = NULL,
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  questID = NULL,
  version = NULL
)
}
\arguments{
\item{df}{dataframe with upload data}

\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{the questionnaire id}

\item{version}{the questionnaire version}
}
\description{
Creates assignment with prefilled data. Uses the httr2 package,
and specifically the \code{\link[httr2]{req_perform_parallel}} function
}
\details{
Dataframe needs to be provided with columns
for prefilling data, as well as \emph{Quantity} and \emph{ResponsibleName}. Return value is a data.table,
with the prefilling data, if successful
}
