% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_user.R
\name{suso_getINT}
\alias{suso_getINT}
\title{Survey Solutions API call for list of interviewers}
\usage{
suso_getINT(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  sv_id = NULL,
  bigTeam = FALSE
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}

\item{sv_id}{supervisor id, if NULL all interviewers in the workspace are returned, and a column with sv_id is added}

\item{bigTeam}{(only if sv_id=NULL) if TRUE, requests will be performed sequential, and individual teams can be bigger than 100 if FALSE
requests will be performed in parallel, but do not allow for teams larger 100 interviewers.}
}
\description{
Get list of all interviewers by supervisor id, or a list of all interviewers in the workspace.
}
\examples{
\dontrun{

# Get all members for a single supervisor
suso_getINT(
          sv_id = "xxxx-xxxx-xxxx-xxx"
          )

# Get all members in the workspace
suso_getINT()

}

}
