% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_user.R
\name{suso_getINT}
\alias{suso_getINT}
\title{Survey Solutions API call for list of interviewers}
\usage{
suso_getINT(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  sv_id = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}

\item{sv_id}{supervisor id, if NULL all interviewers in the workspace are returned}
}
\description{
Get list of all interviewers by supervisor id, or a list of all interviewers in the workspace.
}
\examples{
\dontrun{
suso_getINT(
          workspace = "myworkspace",
          sv_id = "xxxx-xxxx-xxxx-xxx"
          )
}

}
