% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_Interviews.R
\name{suso_patchRejectInterview}
\alias{suso_patchRejectInterview}
\title{Reject interviews either as supervisor or as headquarter.}
\usage{
suso_patchRejectInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = suso_get_api_key("workspace"),
  token = NULL,
  intID = "",
  new_uid = NULL,
  HQ = FALSE,
  comment = "Please check errors and re-submit!"
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{intID}{the \emph{InterviewId} of the interview.}

\item{new_uid}{if provided the interview will be rejected to a different user than the current one.}

\item{HQ}{if FALSE, reject as supervisor, if TRUE rejected as headquarters}

\item{comment}{comment which should be sent with the questionnaire}
}
\description{
Allows you to reject interviews in supervisor or headquarters role
as well as to provide a comment (i.e. reason) for the rejection.
}
\details{
For details please see:
\url{https://docs.mysurvey.solutions/headquarters/interviews/survey-workflow/}
}
\examples{
\dontrun{
# reject the interview as supervisor
suso_patchRejectInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea"
          )
# reject the interview as headquarters
suso_patchRejectInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
          HQ = TRUE
          )
# reject the interview and provide a comment, so the interviewer knows about the problem
suso_patchRejectInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
          comment = "Too many errors, please check and re-submit!"
          )
# reject the interview to a different user
suso_patchRejectInterview(
          workspace = "myworkspace",
          intID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
          new_uid = "3b0c6e09-d606-4914-9e20-2abc048d5bea"
          )

}

}
