% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_suso_api_credentials.R
\name{suso_set_workspace}
\alias{suso_set_workspace}
\title{Convenience function to switch workspace}
\usage{
suso_set_workspace(workspace = NULL)
}
\arguments{
\item{workspace}{server workspace Name (not the display name), if nothing provided, defaults to primary}
}
\value{
invisibly TRUE if successful.
}
\description{
Sets the workspace only, but leaves all other credentials the same.
}
\details{
Use \code{suso_set_workspace} to make the desired workspace available for all the \code{suso_}
functions, so you don't need to specify the workspace parameter within those
functions. The function also checks if the workspace name is correct, and the user with the current credentials is
authorized.If the workspace requires different credentials, then use \code{\link{suso_set_key}} again.
}
\examples{
\dontrun{

# switch to workspace "windows"
suso_set_workspace("windows")

# switch to primary (default) workspace
suso_set_workspace()

}

}
